chore(pointers): move classes into separate files
